DataStore.java is changed to BerkeleyDBDataStore.java.
A new datastore interface is added to jessy called DataStore.java.

A new data store called hashmapDataStore.java is added to jessy. It is a simple in-memory data store using java concurrent hashmap impl. Some of the not-required methods for the moment are not imlemented.

Selecting a data store is provided via DataStoreFactory.java class. The parameter should be set inside config.property file.

Test cases for data store are changed accordingly, and now support both of the implementations.

Consistency implementations now use the DataStore interface.