The problem in the voting phase is fixed. Previously, the voting phase for US/SER/SI was implemented wrongly. This patch fixes the problem for SI. However, US/SER are not working properly. Needs further investigation. Moreover, some problems regarding HashMapDataStore and ScalarVector is addressed.
Details:
-With this patch, the set of replicas that send the votes to the
coordinator can be defined manually (needed for SI).
-There is a distinction between vote receivers and vote senders. For
each consistency, the set of vote senders and vote receivers are defined
manually.
-SI voting phase is fixed and working.
-SER/US voting phase is fixed, but not working.
-HashmapDataStore now efficiently handles NEVER_COMPTIBLE comparator
returned from ScalarVector. I.e., upon receiving a NEVER_COMPATIBLE and
instead sending null back to the client, so the client asks another
replica, it retries reading a new version again.
-HashMapDataStore now clones a read entity before checking its
compatibility. This is required for SI/PSI consistency since the local
vector of the entity can be changed inside isCompatible method. (TODO:
for the sake of performance, this patch should only works for SI/PSI).
-ScalarVector is now working perfectly. Upon receiving a read request
for a key such that the local GSN is smaller than the transaction
snapshot number, the reading thread sleeps in the scalar vector until
the read can be performed. It then returns NEVER_COMPATIBLE, and the
data store can either return null to the client or just try to read
again. This time it will surely read.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/regal/trunk/src/jessy@2083 334080fa-30e0-4884-a663-39494f8d70c4