Several issues regarding DistributedTermination are resolved.
Details:
1. Convoy effect handling in Distributed Termination was not optimized
because it was ordering all transactions before certification. The
algorithm is changed such that now transactions run concurrently (for
certification and vote exchanging) in the jessy replicas as long as
there is no consisitency conflicts. However, after certification and
voting exchange, they are queued for being applied to the data store
only on repicas replicating objects modified inside the transactions.
Otherwise, they proceed again. This impacts the performance alot.

2. Different implementation of different consistencies (SI, US, NMSI,
SER) are modified according to OO rules. e.g., there is an abstract NMSI
class, and two concrete implementation of it based on GMUVector and DV.

4. The problem regarding the measurement of certification latency is
fixed. There problem stemed from the fact that certification latency was
being computed for both the loading and execution phases. Now, it is
reset after the loading phase and at the begining of execution phase.

3. Some probes were added to DistributedTransaction for measuring
different parameters in the jessy replicas.

In addition, some of the useless probes were removed from
HashMapDataStore.