Bugs that are observed under high update ratio are addressed. Before performing a unicast, the channel is now checked to make sure that it is still connected. If the connection of a channel is broken, it tries to reconnect before performing the unicast.
The way Votes are handled is now changed. Each jessy replica first sends its vote to other jessy replicas, and then sends it to the jessy proxy.
This is because we observed that under high update ratio, the performance of the system was not good. The reason is because if we send the vote
back to the proxy right away, the client will re-execute the transaction right away while the replicas have not garbage collected the aborted transaction.
Thus leads to high loads at the jessy replicas.

If TerminationTransactionRequestMessage is the result of an aborted transaction, it also piggy backs the transaction uuid of that aborted transaction.
This is necessary to rapidly remove the aborted transactions from the servers. Otherwise, it might take some time to garbage collect them, and all resulting transactions
must wait in the queue because they have conflicts with the aborted yet existed transaction.

The probes for calculating Execution and Termination latency in the Transaction class are fixed.

A new class is created for applying the transactions sequentiall. This results in substantial performance improvement compare with the previous method of waiting in the TransactionTermination task.In this approach, we add the transaction to a queue, and release the thread right away. There is only one thread that sequentially applies all the transactions to the data store.

Some timeouts in the constant pool are changed.
Some warning messages were removed.