Problems regarding PSI when group size is greater than 1 are fixed as follow:
1. Upon being certified in a jessy replica, when group size is greater than 1, transactions commits together.
This leads to having several different transactions in different members of a group with different sequence number.
The fix is that certification commutativity test always returns false if group size is greater than 1.

2. Size of ParallelSnapshotIsolationPiggyback is reduced by removing the ExecutionHistory from it
and only storing needed data.

Note, there is a dirty quick (not safe) fix as well, that should be addressed properly. (issue #13)