Now supports read/write to disk for bypassing loading phase of YCSB.
Issues regarding concurrent certifying transactions are resolved when the group size is greater than 1.
Previously, this problem was resolved for only PSI, but NMSI and US also suffer this problem,
and the fix for PSI was not good either. So, a new algorithm is implemented in DistributedTermination
for handling concurrency when group size is greater than one.