LightGenuineTerminationCommuinication is implemented. This implementation only am-cast transaction handler, and TerminateTransactionRequestMessage is rm-cast. This improves the latency of am-cast by 70%.
TerminationCommunication is now generic. There are several different implementation of it. Moreover, VoteMulticast is made generic for better usability.

A TerminationCommunicationFactory is implemented for returning a required TerminationCommunication instance based on consistency and user preference.
All consistencies are modified accordingly.

DistributedTermination is modified and a unicast manager for receiving votes is removed from it. It is moved to VoteMulticast primitive.