Fix P-Store Fix naming convention for protocols. Add back getVector method in VectorFactory.
P-Store was unable to commit an update transaction because during certify test, it did not check whether the object is replicated locally or not. This caused null pointer exception, and was fixed.
Some vectors need to be initialize during create time. Previously, it was the case, but at some point, the method was removed. Since loaded DB was always used after that change, no problem ever occured. But with changes to JessyEntity, and the need to load DB again from scratch, an exception happened always during loading time. This problem was fixed.