[FIX] the bug in YCSB benchmark regarding number of transactions per thread.
In the original YCSB, the number of transactions (i.e., operations) are devided by the number of threads.
The result is passed as an integer.
Hence, if the number of transactions are not devidable by the number of threads, then YCSB executes less transactions than defined in the configuration.
To fix this, the remainder of the division is added to the first threads that are being created.